+++ src/cli.rs
@@ -13,7 +13,7 @@ use clap::{Args, Parser, Subcommand};
 use std::convert::{TryFrom, TryInto};
 use std::ffi::{OsStr, OsString};
 
-pub(in crate::cli) mod configurators;
+pub mod configurators;
 pub(crate) mod custom_check_opts;
 pub(crate) mod find_opts;
 pub(crate) mod rust_releases_opts;
@@ -66,7 +66,7 @@ fn modify_args<I: IntoIterator<Item = T>, T: Into<OsString> + Clone>(
 }
 
 #[derive(Debug, Subcommand)]
-pub(in crate::cli) enum CargoMsrvCli {
+pub enum CargoMsrvCli {
     /// Find your Minimum Supported Rust Version!
     #[command(
         author = "Martijn Gribnau <garm@ilumeo.com>",
@@ -88,24 +88,20 @@ pub(in crate::cli) enum CargoMsrvCli {
 
 #[derive(Debug, Args)]
 #[command(version)]
-pub(in crate::cli) struct CargoMsrvOpts {
+pub struct CargoMsrvOpts {
     #[command(flatten)]
-    pub(in crate::cli) find_opts: FindOpts,
+    pub find_opts: FindOpts,
 
     #[command(flatten)]
-    pub(in crate::cli) shared_opts: SharedOpts,
+    pub shared_opts: SharedOpts,
 
     #[command(subcommand)]
-    pub(in crate::cli) subcommand: Option<SubCommand>,
-
-    /// DEPRECATED: Use the `cargo msrv verify` subcommand instead
-    #[arg(long, global = false, hide = true)]
-    pub(in crate::cli) verify: bool,
+    pub subcommand: Option<SubCommand>,
 }
 
 #[derive(Debug, Subcommand)]
 #[command(propagate_version = true)]
-pub(in crate::cli) enum SubCommand {
+pub enum SubCommand {
     /// Display the MSRV's of dependencies
     List(ListOpts),
     /// Set the MSRV of the current crate to a given Rust version
@@ -119,15 +115,15 @@ pub(in crate::cli) enum SubCommand {
 
 #[derive(Debug, Args)]
 #[command(next_help_heading = "List options")]
-pub(in crate::cli) struct ListOpts {
+pub struct ListOpts {
     /// Display the MSRV's of crates that your crate depends on
     #[arg(long, value_enum, default_value_t)]
-    variant: ListMsrvVariant,
+    pub variant: ListMsrvVariant,
 }
 
 #[derive(Debug, Args)]
 #[command(next_help_heading = "Set options")]
-pub(in crate::cli) struct SetOpts {
+pub struct SetOpts {
     /// The version to be set as MSRV
     ///
     /// The given version must be a two- or three component Rust version number.
@@ -135,26 +131,26 @@ pub(in crate::cli) struct SetOpts {
     /// in the Cargo manifest. MSRV's greater or equal to 1.56 will be written to
     /// `package.rust-version` in the Cargo manifest.
     #[arg(value_name = "MSRV")]
-    msrv: BareVersion,
+    pub msrv: BareVersion,
 }
 
 #[derive(Debug, Args)]
 #[command(next_help_heading = "Verify options")]
-pub(in crate::cli) struct VerifyOpts {
+pub struct VerifyOpts {
     #[command(flatten)]
-    pub(in crate::cli) rust_releases_opts: RustReleasesOpts,
+    pub rust_releases_opts: RustReleasesOpts,
 
     #[command(flatten)]
-    pub(in crate::cli) toolchain_opts: ToolchainOpts,
+    pub toolchain_opts: ToolchainOpts,
 
     #[command(flatten)]
-    pub(in crate::cli) custom_check: CustomCheckOpts,
+    pub custom_check: CustomCheckOpts,
 
     /// The Rust version, to check against for toolchain compatibility
     ///
     /// If not set, the MSRV will be parsed from the Cargo manifest instead.
     #[arg(long, value_name = "rust-version")]
-    rust_version: Option<BareVersion>,
+    pub rust_version: Option<BareVersion>,
 }
 
 // Interpret the CLI config frontend as general Config
@@ -216,13 +212,7 @@ fn make_mode(opts: &CargoMsrvOpts) -> SubcommandId {
             SubCommand::Set(_) => SubcommandId::Set,
             SubCommand::Verify(_) => SubcommandId::Verify,
         })
-        .unwrap_or_else(|| {
-            if opts.verify {
-                SubcommandId::Verify
-            } else {
-                SubcommandId::Find
-            }
-        })
+        .unwrap_or_else(|| SubcommandId::Find)
 }
 
 #[cfg(test)]